+++ crates/router/src/core/webhooks/recovery_incoming.rs
@@ -1461,7 +1461,7 @@ impl RecoveryPaymentTuple {
     pub async fn publish_revenue_recovery_event_to_kafka(
         state: &SessionState,
         recovery_payment_tuple: &Self,
-        retry_count: Option<i32>,
+        retry_count: Option<i64>,
     ) -> CustomResult<(), errors::RevenueRecoveryError> {
         let recovery_payment_intent = &recovery_payment_tuple.0;
         let recovery_payment_attempt = &recovery_payment_tuple.1;
@@ -1499,7 +1499,7 @@ impl RecoveryPaymentTuple {
         #[allow(clippy::as_conversions)]
         let retry_count = Some(retry_count.unwrap_or_else(|| {
             revenue_recovery_feature_metadata
-                .map(|data| data.total_retry_count as i32)
+                .map(|data| data.total_retry_count as i64)
                 .unwrap_or(0)
         }));
