+++ crates/router/src/workflows/payment_sync.rs
@@ -273,7 +273,7 @@ pub async fn get_sync_process_schedule_time(
     db: &dyn StorageInterface,
     superposition_client: &external_services::superposition::SuperpositionClient,
     dimensions: &dimension_state::DimensionsWithProcessorMerchantIdAndConnector,
-    retry_count: i32,
+    retry_count: i64,
 ) -> Result<Option<time::PrimitiveDateTime>, errors::ProcessTrackerError> {
     let mapping = dimensions
         .get_pt_mapping_payment_sync(db, superposition_client, None)
